Create SolutionAVeryBigSum
We can calculate the sum of the elements in the array by traversing the array once, summing each element as we go. Because this sum can be quite large, we must use a long data type to hold the value of our sum. A signed −bit integer uses the first bit to represent the sign of the number and the remaining  bits to represent the magnitude. The range of a 32−bit integer is  -2 ^31 to 2^31 -1 , or[-2147483648, 2147483647] . When we add several integer values, the resulting sum might exceed this range. Use long long in C/C++ and long in Java.